requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=8.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
opencv-python>=4.5.0
mtcnn>=0.1.1
